Create history module and move testing functions from test_linkedList into a new module
The new testing module can now be used by all program modules so that we
don't have to rewrite the utilities.
