+++ history/Makefile
@@ -1,0 +1,26 @@
+CC= gcc
+CFLAGS= -Wall -g -pedantic -std=c99 
+LDFLAGS= -lm
+
+test_history: history.o test_history.o ../linkedList/linkedList.o
+	$(CC) $(CFLAGS) $^ -o $@
+
+test_history.o: test_history.c 
+	$(CC) $(CFLAGS) -c test_history.c
+
+history.o: history.c history.h
+	$(CC) $(CFLAGS) -c history.c
+
+linkedList.o:
+	make -C ../linkedList/
+
+.PHONY:
+
+test: memCheck
+	./test_history
+
+memCheck: test_history
+	valgrind -s --leak-check=yes ./test_history
+
+clean:
+	rm -f *~ *.o test_history
