+++ test/Makefile
@@ -1,0 +1,9 @@
+CC= gcc
+CFLAGS= -Wall -g -pedantic -std=c99 
+
+testUtils.o: testUtils.c
+	$(CC) $(CFLAGS) -c testUtils.c
+
+.PHONY:
+clean:
+	rm -f *~ *.o
